backend/requirements.txt
fastapi
uvicorn
python-multipart

pandas
numpy
scikit-learn
scipy

google-generativeai
python-dotenv
langchain
langchain-google-genai

fpdf
openpyxl